+++ src/cube/edges.h
@@ -11,22 +11,22 @@ public:
     std::vector<int> edges_orient;
     void apply_move(int move);
     std::pair<uint32_t, uint32_t> to_index() const;
-    void to_array(uint32_t state);
+    void to_array(uint64_t state);
 private:
     int pieces_cnt;
     std::array<Table12, 6> perm, orient;
-    std::vector<int> offsets;
+    std::vector<uint32_t> offsets;
 };
 
-constexpr int product(int a, int b)
+constexpr uint64_t product(int a, int b)
 {
-    int res = 1;
+    uint64_t res = 1;
     for (int x = a; x <= b; x++)
         res *= x;
     return res;
 }
 
-constexpr int factorial(int x)
+constexpr uint64_t factorial(int x)
 {
     return product(1, x);
 }
