+++ src/pruning/edgePruning.h
@@ -8,7 +8,7 @@ public:
     edgePruning(int pieces_cnt);
     int pruning_number(Edges &edges);
     int pruning_number(Cube &cube) { return 0; }
-    int visited;
+    uint64_t visited;
 private:
     void buildPruneTable();
     void pruneTreeSearch(Edges & edges, char depth_left, char depth, int lastMove);
